+++ .github/workflows/ci.yaml
@@ -19,7 +19,7 @@ jobs:
 
     strategy:
       matrix:
-        targetplatform: [x64]
+        targetplatform: [x86]
 
     runs-on: windows-latest
