--- Daybreak.Injector/Daybreak.Injector.csproj
@@ -1,19 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>net8.0</TargetFramework>
-    <RootNamespace>Daybreak.Injector</RootNamespace>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-    <Platforms>AnyCPU;x86</Platforms>
-    <PublishSingleFile>True</PublishSingleFile>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="8.0.0" />
-    <PackageReference Include="SystemExtensions.NetCore" Version="1.0.1" />
-    <PackageReference Include="SystemExtensions.NetStandard" Version="1.6.2" />
-  </ItemGroup>
-
-</Project>
